feat: now we can add a few homeTypes and add properties to form

+++ src/main/kotlin/souliving/backend/dto/PlainShortFormDto.kt
@@ -17,7 +17,7 @@ data class PlainShortFormDto(
     var alcohol: Boolean,
     var petfriendly: Boolean?,
     var isclean: Boolean?,
-    var homeownerid: Int? = null,
+    var homeownerid: Long? = null,
     var photoid: Long? = null,
     var onlinedatetime: LocalDateTime?,
 )
